% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGrads-package.R
\docType{data}
\name{ctlparams}
\alias{ctlparams}
\title{Example of a parsed .ctl file}
\description{
Example of a parsed .ctl file
}
\details{
This dataset shows the result of the function \code{\link{parseCTLfile}},
and is used in this package as an example dataset.
}
\author{
Paul Hiemstra, \email{p.h.hiemstra@gmail.com}
}
\keyword{data}
