\name{gradsData2array}
\alias{gradsData2array}
\title{Plots a list of SpatialGrid/PixelDataFrames using ggplot}
\usage{
  gradsData2array(ts, lev, var, gradsData, ctlparams)
}
\arguments{
  \item{a}{b}
}
\value{
  something
}
\description{
  Test 1
}
\examples{
print("Hello World")
}
\author{
  Paul Hiemstra, \email{p.h.hiemstra@gmail.com}
}

