% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGrads-package.R
\docType{data}
\name{gradsExampleData}
\alias{gradsExampleData}
\title{Example of a grads dataset}
\description{
Example of a grads dataset
}
\details{
This dataset shows the result of the function \code{\link{readGradsFile}},
and is used in this package as an example dataset.
}
\author{
Paul Hiemstra, \email{p.h.hiemstra@gmail.com}
}
\keyword{data}
