% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridt.R
\name{gridt}
\alias{gridt}
\title{Extract date and time information from content of ctl file}
\usage{
gridt(ntimes, time0, dtime, century = 2000)
}
\arguments{
\item{ntimes}{numeric}

\item{time0}{character}

\item{dtime}{character}

\item{century}{numeric, default is 2000}
}
\value{
character
}
\description{
Extract date and time information from content of ctl file
}
\author{
Marcos Longo
}
