\name{readListGradsFiles2array}
\alias{readListGradsFiles2array}
\title{Plots a list of SpatialGrid/PixelDataFrames using ggplot}
\usage{
  readListGradsFiles2array(filelist, ts, lev, var, verbose)
}
\arguments{
  \item{a}{b}
}
\value{
  something
}
\description{
  Test 1
}
\examples{
print("Hello World")
}
\author{
  Paul Hiemstra, \email{p.h.hiemstra@gmail.com}
}

