\name{subsetGradsData}
\alias{subsetGradsData}
\title{Plots a list of SpatialGrid/PixelDataFrames using ggplot}
\usage{
  subsetGradsData(ts, lev, var, gradsData, ctlparams,
  correctionFactor = 1)
}
\arguments{
  \item{a}{b}
}
\value{
  something
}
\description{
  Test 1
}
\examples{
print("Hello World")
}
\author{
  Paul Hiemstra, \email{p.h.hiemstra@gmail.com}
}

