\name{gradsData2array}
\alias{gradsData2array}
\title{Dumps subsets of grads data into a data.frame}
\usage{
  gradsData2array(ts, lev, var, gradsData, ctlparams)
}
\arguments{
  \item{ts}{numeric vector giving the timesteps which are
  to be included, e.g. 1:100.}

  \item{lev}{numeric vector expressing which levels are to
  be included, e.g. lev = c(200, 500).}

  \item{var}{character vector expressing which variable(s)
  are to be included, e.g. var = "psi". Note that there is
  currently no support for converting more than one var at
  a time.}

  \item{gradsData}{the grads data, object is generated by
  \code{\link{readGradsFile}}.}

  \item{ctlparams}{the parameters of the grads data, object
  is generated by \code{\link{parseCTLfile}}.}
}
\value{
  a multi-dimensional array with the requested subset.
}
\description{
  This funcion takes the output of
  \code{\link{readGradsFile}} and converts it to an array.
  The function allows the user to take subsets in terms of
  timesteps, variables and levels.
}
\examples{
data(ctlparams)
data(gradsExampleData)

dat1 = gradsData2array(1, 500, "u", gradsExampleData, ctlparams)
dat2 = gradsData2array(1:2, 200, c("u"), gradsExampleData, ctlparams)
dat3 = gradsData2array(1:2, 800, "psi", gradsExampleData, ctlparams)
}
\author{
  Paul Hiemstra, \email{p.h.hiemstra@gmail.com}
}

