\name{parseCTLfile}
\alias{parseCTLfile}
\title{Parses a grads .ctl file into an R list}
\usage{
  parseCTLfile(ctlfname)
}
\arguments{
  \item{ctlfname}{path where the .ctl file is located}
}
\value{
  R list containing the parsed parameters.
}
\description{
  Parses the information in a .ctl file into an R list for
  other readGrads functions to use.
}
\examples{
print("Hello World")
}
\author{
  Paul Hiemstra, \email{p.h.hiemstra@gmail.com}
}

