\name{readGradsFile}
\alias{readGradsFile}
\title{Plots a list of SpatialGrid/PixelDataFrames using ggplot}
\usage{
  readGradsFile(gradsfile, ctlparams, tstepRange,
  convert2dataframe = FALSE, padding.bytes = TRUE)
}
\arguments{
  \item{a}{b}
}
\value{
  something
}
\description{
  Test 1
}
\examples{
print("Hello World")
}
\author{
  Paul Hiemstra, \email{p.h.hiemstra@gmail.com}
}

