% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readListGradsFiles2dataframe.R
\name{readListGradsFiles2dataframe}
\alias{readListGradsFiles2dataframe}
\title{Reads a list of grads files into a data.frame}
\usage{
readListGradsFiles2dataframe(filelist, ts, lev, var, verbose)
}
\arguments{
\item{filelist}{character vector providing the locations of the grads files
are to be read.}

\item{ts}{numeric vector expressing which timesteps should be read, e.g. ts = 1:10.}

\item{lev}{numeric vector expressing which levels should be read, e.g. lev = c(200, 500).}

\item{var}{character vector expressing which variable(s) should be read, e.g. var = "psi".}

\item{verbose}{logical, if TRUE, the function provides feedback on what it is doing.}
}
\value{
a data.frame with the requested data.
}
\description{
Reads a list of grads files into an array and then
uses melt to transform the array to a data.frame
}
\author{
Paul Hiemstra, \email{p.h.hiemstra@gmail.com}
}
