\name{readListGradsFiles2dataframe}
\alias{readListGradsFiles2dataframe}
\title{Reads a list of grads files into a data.frame}
\description{
  Reads a list of grads files into a data.frame
}
\note{
  This function is not yet implemented
}
\examples{
print("Hello World")
}
\author{
  Paul Hiemstra, \email{p.h.hiemstra@gmail.com}
}

