\name{subsetGradsData}
\alias{subsetGradsData}
\title{Dumps subsets of grads data into a data...}
\usage{subsetGradsData(ts, lev, var, gradsData, ctlparams, correctionFactor=1)}
\description{Dumps subsets of grads data into a data.frame}
\details{This funcion takes the output of \code{\link{readGradsFile}} and
converts it to a data.frame. The function allows the user to 
take subsets in terms of timesteps, variables and levels.}
\value{something}
\author{Paul Hiemstra, \email{p.h.hiemstra@gmail.com}}
\arguments{\item{ts}{numeric vector giving the timesteps which are to be included, e.g. 1:100.}
\item{lev}{numeric vector expressing which levels are to be included, e.g. lev = c(200, 500).}
\item{var}{character vector expressing which variable(s) are to be included, e.g. var = "psi".}
\item{gradsData}{the grads data, object is generated by \code{\link{readGradsFile}}.}
\item{ctlparams}{the parameters of the grads data, object is generated by \code{\link{parseCTLfile}}.}
\item{correctionFactor}{the values read from the grads files are multiplied by this number.}}
\examples{data(ctlparams)
data(gradsExampleData)

dat1 = subsetGradsData(1, 500, "u", gradsExampleData, ctlparams)
dat2 = subsetGradsData(1:2, 200, c("u","v"), gradsExampleData, ctlparams)
dat3 = subsetGradsData(1:2, 800, "psi", gradsExampleData, ctlparams)}
